+++ includes/class-tmsm-woocommerce-paymentonsite-gateway.php
@@ -5,6 +5,7 @@
  * @package WooCommerce\Gateways
  */
 
+use Automattic\Jetpack\Constants;
 
 if ( ! defined( 'ABSPATH' ) ) {
 	exit; // Exit if accessed directly.
@@ -59,15 +60,13 @@ class WC_Gateway_Paymentonsite extends WC_Gateway_COD {
 	 * Override Gateway Settings Form Fields.
 	 */
 	public function override_form_fields() {
-
 		$this->form_fields['enabled']['label'] = __( 'Enable payment on site', 'tmsm-woocommerce-paymentonsite-status' );
 		$this->form_fields['title']['default'] = __( 'Payment on site', 'tmsm-woocommerce-paymentonsite-status' );
 		$this->form_fields['description']['default'] = __( 'Payment on site', 'tmsm-woocommerce-paymentonsite-status' );
 		$this->form_fields['instructions']['description'] = __( 'Instructions that will be added to the thank you page and confirmation email. Use the placeholder {shop_address} for the shop address.', 'tmsm-woocommerce-paymentonsite-status' );
 		$this->form_fields['instructions']['default'] = __( 'Payment on site at {shop_address}', 'tmsm-woocommerce-paymentonsite-status' );
-		//$this->form_fields['enable_for_methods']['options'] = $options;
+		$this->form_fields['enable_for_methods']['options'] = $this->load_shipping_method_options();
 		$this->form_fields['enable_for_virtual']['label'] = __( 'Accept Payment On Site if the order is virtual', 'tmsm-woocommerce-paymentonsite-status' );
-
 	}
 
 	/**
@@ -187,4 +186,95 @@ class WC_Gateway_Paymentonsite extends WC_Gateway_COD {
 		}
 		return $appointmentonly;
 	}
+
+	/**
+	 * Checks to see whether or not the admin settings are being accessed by the current request.
+	 *
+	 * @return bool
+	 */
+	private function is_accessing_settings() {
+
+		if ( is_admin() ) {
+			// phpcs:disable WordPress.Security.NonceVerification
+			if ( ! isset( $_REQUEST['page'] ) || 'wc-settings' !== $_REQUEST['page'] ) {
+				return false;
+			}
+			if ( ! isset( $_REQUEST['tab'] ) || 'checkout' !== $_REQUEST['tab'] ) {
+				return false;
+			}
+			if ( ! isset( $_REQUEST['section'] ) || 'paymentonsite' !== $_REQUEST['section'] ) {
+				return false;
+			}
+			// phpcs:enable WordPress.Security.NonceVerification
+
+			return true;
+		}
+
+		if ( Constants::is_true( 'REST_REQUEST' ) ) {
+			global $wp;
+			if ( isset( $wp->query_vars['rest_route'] ) && false !== strpos( $wp->query_vars['rest_route'], '/payment_gateways' ) ) {
+				return true;
+			}
+		}
+
+		return false;
+	}
+
+	/**
+	 * Loads all of the shipping method options for the enable_for_methods field.
+	 *
+	 * @return array
+	 */
+	private function load_shipping_method_options() {
+
+		echo '***paymentonsite load_shipping_method_options';
+		// Since this is expensive, we only want to do it if we're actually on the settings page.
+		if ( ! $this->is_accessing_settings() ) {
+			return array();
+		}
+		echo '***paymentonsite load_shipping_method_options ok';
+
+		$data_store = WC_Data_Store::load( 'shipping-zone' );
+		$raw_zones  = $data_store->get_zones();
+
+		foreach ( $raw_zones as $raw_zone ) {
+			$zones[] = new WC_Shipping_Zone( $raw_zone );
+		}
+
+		$zones[] = new WC_Shipping_Zone( 0 );
+
+		$options = array();
+		foreach ( WC()->shipping()->load_shipping_methods() as $method ) {
+
+			$options[ $method->get_method_title() ] = array();
+
+			// Translators: %1$s shipping method name.
+			$options[ $method->get_method_title() ][ $method->id ] = sprintf( __( 'Any &quot;%1$s&quot; method', 'woocommerce' ), $method->get_method_title() );
+
+			foreach ( $zones as $zone ) {
+
+				$shipping_method_instances = $zone->get_shipping_methods();
+
+				foreach ( $shipping_method_instances as $shipping_method_instance_id => $shipping_method_instance ) {
+
+					if ( $shipping_method_instance->id !== $method->id ) {
+						continue;
+					}
+
+					$option_id = $shipping_method_instance->get_rate_id();
+
+					// Translators: %1$s shipping method title, %2$s shipping method id.
+					$option_instance_title = sprintf( __( '%1$s (#%2$s)', 'woocommerce' ), $shipping_method_instance->get_title(), $shipping_method_instance_id );
+
+					// Translators: %1$s zone name, %2$s shipping method instance name.
+					$option_title = sprintf( __( '%1$s &ndash; %2$s', 'woocommerce' ), $zone->get_id() ? $zone->get_zone_name() : __( 'Other locations', 'woocommerce' ), $option_instance_title );
+
+					$options[ $method->get_method_title() ][ $option_id ] = $option_title;
+				}
+			}
+		}
+
+		return $options;
+	}
+
 }